+++ ByteFlood/App.xaml
@@ -18,7 +18,7 @@
     <Application x:Class="ByteFlood.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             StartupUri="MainWindow.xaml">
+             StartupUri="UI\MainWindow.xaml">
     <Application.Resources>
         <ResourceDictionary>
         </ResourceDictionary>
