+++ ByteFlood/ByteFlood.csproj
@@ -38,6 +38,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="PresentationFramework" />
+    <Reference Include="PresentationFramework.Aero" />
     <Reference Include="PresentationFramework.Aero2, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
       <SpecificVersion>False</SpecificVersion>
       <HintPath>..\..\..\..\..\..\..\Windows\Microsoft.NET\Framework\v4.0.30319\WPF\PresentationFramework.Aero2.dll</HintPath>
@@ -62,12 +63,20 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Compile Include="AddTorrentDialog.xaml.cs">
+    <Compile Include="InfoClasses\FileInfo.cs" />
+    <Compile Include="InfoClasses\PeerInfo.cs" />
+    <Compile Include="InfoClasses\PieceInfo.cs" />
+    <Compile Include="InfoClasses\TrackerInfo.cs" />
+    <Compile Include="UI\AddTorrentDialog.xaml.cs">
       <DependentUpon>AddTorrentDialog.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Formatters\HMSFormatter.cs" />
+    <Compile Include="Formatters\SizeToString.cs" />
+    <Compile Include="Formatters\SpeedFormatter.cs" />
+    <Compile Include="Formatters\TimeSpanToString.cs" />
     <Compile Include="GraphDrawer.cs" />
-    <Compile Include="InfoClasses.cs" />
-    <Compile Include="Preferences.xaml.cs">
+    <Compile Include="InfoClasses\TorrentInfo.cs" />
+    <Compile Include="UI\Preferences.xaml.cs">
       <DependentUpon>Preferences.xaml</DependentUpon>
     </Compile>
     <Compile Include="Properties\Resources.Designer.cs">
@@ -76,14 +85,15 @@
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
     <Compile Include="Settings.cs" />
-    <Compile Include="TorrentProperties.xaml.cs">
+    <Compile Include="UI\TorrentProperties.xaml.cs">
       <DependentUpon>TorrentProperties.xaml</DependentUpon>
     </Compile>
-    <Page Include="AddTorrentDialog.xaml">
+    <Compile Include="Utility.cs" />
+    <Page Include="UI\AddTorrentDialog.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="MainWindow.xaml">
+    <Page Include="UI\MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
@@ -91,15 +101,15 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="MainWindow.xaml.cs">
+    <Compile Include="UI\MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Page Include="Preferences.xaml">
+    <Page Include="UI\Preferences.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="TorrentProperties.xaml">
+    <Page Include="UI\TorrentProperties.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -137,6 +147,7 @@
       <Name>MonoTorrent</Name>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
