+++ ByteFlood/UI/AddTorrentDialog.xaml
@@ -1,0 +1,71 @@
+﻿<!-- 
+    ByteFlood - A BitTorrent client.
+    Copyright (C) 2014 Burak Öztunç
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+    -->
+    <Window x:Class="ByteFlood.AddTorrentDialog"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        Title="AddTorrent" Height="352" Width="723" Loaded="Window_Loaded">
+    <Grid>
+        <GroupBox Header="Save in" Height="73" HorizontalAlignment="Left" Margin="13,0,0,0" Name="gb1" VerticalAlignment="Top" Width="324">
+            <Grid Margin="2,0,-2,0">
+                <TextBox Height="23" Text="" HorizontalAlignment="Left" Margin="6,14,0,0" Name="pathbox" VerticalAlignment="Top" Width="251" />
+                <Button Content="..." Height="23" HorizontalAlignment="Left" Margin="263,14,0,0" Name="select" VerticalAlignment="Top" Width="43" Click="button1_Click" />
+            </Grid>
+        </GroupBox>
+        <GroupBox Header="Contents" HorizontalAlignment="Left" Margin="342,0,0,0" Name="gb2" Width="347" VerticalAlignment="Top" Height="274">
+            <Grid>
+                <Label Content="Name:" Height="28" HorizontalAlignment="Left" Name="label1" VerticalAlignment="Top" Margin="6,6,0,0" />
+                <Label Content="Comment:" Height="28" HorizontalAlignment="Left" Margin="6,29,0,0" Name="label2" VerticalAlignment="Top" />
+                <Label Content="Size:" Height="28" HorizontalAlignment="Left" Margin="6,52,0,0" Name="label4" VerticalAlignment="Top" />
+                <Label Content="Date:" Height="28" HorizontalAlignment="Left" Margin="6,75,0,0" Name="label5" VerticalAlignment="Top" />
+                <Label Content="{Binding Torrent.Name}" Height="28" HorizontalAlignment="Left" Margin="76,6,0,0" Name="label3" VerticalAlignment="Top" />
+                <Label Content="{Binding Torrent.Comment}" Height="28" HorizontalAlignment="Left" Margin="76,29,0,0" Name="label6" VerticalAlignment="Top" />
+                <Label Height="28" HorizontalAlignment="Left" Margin="76,52,0,0" Name="size" VerticalAlignment="Top" />
+                <Label Content="{Binding Torrent.CreationDate}" Height="28" HorizontalAlignment="Left" Margin="76,75,0,0" Name="label8" VerticalAlignment="Top" />
+                <ListView HorizontalAlignment="Left" Margin="6,108,0,6" Name="filelist" Width="323" SelectionMode="Single" SelectedIndex="0">
+                    <ListView.View>
+                        <GridView>
+                            <GridViewColumn Width="230" Header="Name" >
+                                <GridViewColumn.CellTemplate>
+                                    <DataTemplate>
+                                        <StackPanel Orientation="Horizontal">
+                                            <CheckBox IsChecked="True" Checked="CheckBox_Checked" Width="Auto" Click="CheckBox_Click" Height="Auto" />
+                                            <TextBlock Text="{Binding Name}" Margin="5,0,0,0" Width="Auto" Height="Auto" />
+                                        </StackPanel>
+                                    </DataTemplate>
+                                </GridViewColumn.CellTemplate>
+                            </GridViewColumn>
+                            <GridViewColumn Header="Size" DisplayMemberBinding="{Binding Size}"  Width="60" />
+                        </GridView>
+                    </ListView.View>
+                </ListView>
+            </Grid>
+        </GroupBox>
+        <GroupBox Header="Torrent options" Height="196" HorizontalAlignment="Left" Margin="13,78,0,0" Name="groupBox3" VerticalAlignment="Top" Width="324">
+            <Grid>
+                <Label Content="Name: " Height="28" HorizontalAlignment="Left" Margin="6,6,0,0" Name="label7" VerticalAlignment="Top" />
+                <TextBox Height="23" HorizontalAlignment="Left" Margin="60,10,0,0" Name="name" VerticalAlignment="Top" Width="246" />
+                <CheckBox Content="Start immediately" Height="16" HorizontalAlignment="Left" Margin="6,40,0,0" Name="start_torrent" IsChecked="True" VerticalAlignment="Top" />
+                <Label Content="Stop seeding when ratio exceeds" Height="28" HorizontalAlignment="Left" Margin="6,62,0,0" Name="label9" VerticalAlignment="Top" />
+                <TextBox Height="23" HorizontalAlignment="Left" Margin="192,65,0,0" Name="ratiolimit" Text="0.000" VerticalAlignment="Top" Width="114" />
+                <CheckBox Content="Initial seeding" Height="16" HorizontalAlignment="Left" IsChecked="True" Margin="212,40,0,0" Name="initial" VerticalAlignment="Top" />
+            </Grid>
+        </GroupBox>
+        <Button Content="OK" Height="23" HorizontalAlignment="Left" Margin="614,280,0,0" Name="button1" VerticalAlignment="Top" Width="75" Click="button1_Click_1" />
+        <Button Content="Cancel" Height="23" HorizontalAlignment="Left" Margin="533,280,0,0" Name="button2" VerticalAlignment="Top" Width="75" Click="button2_Click" />
+    </Grid>
+</Window>
