+++ ByteFlood/UI/Preferences.xaml
@@ -1,0 +1,33 @@
+﻿<Window x:Class="ByteFlood.Preferences"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        Title="Preferences" Height="484" Width="629" Loaded="Window_Loaded">
+    <Grid>
+        <TreeView Height="420" HorizontalAlignment="Left" Margin="12,12,0,0" Name="treeView1" VerticalAlignment="Top" Width="140">
+            <TreeViewItem Header="General" Name="general_tree" IsSelected="True" />
+            <TreeViewItem Header="UI" Name="ui_tree" >
+            </TreeViewItem>
+        </TreeView>
+        <TabControl HorizontalAlignment="Left" Margin="158,10,0,13" Name="tabControl1" Width="437">
+            <TabItem Name="general" IsSelected="{Binding IsSelected, ElementName=general_tree}" Height="0">
+                <Grid>
+                    <CheckBox Content="Prefer encryption" Height="16" HorizontalAlignment="Left" Margin="6,6,0,0" Name="checkBox1" VerticalAlignment="Top" IsChecked="{Binding PreferEncryption}" />
+                    <TextBlock Height="23" HorizontalAlignment="Left" Margin="6,28,0,0" Name="textBlock1" Text="Default download path" VerticalAlignment="Top" />
+                    <TextBox Height="23" HorizontalAlignment="Left" Margin="134,27,0,0" Name="downpath" VerticalAlignment="Top" Width="244" IsReadOnly="True" Text="{Binding DefaultDownloadPath}" />
+                    <Button Content="..." Height="23" HorizontalAlignment="Left" Margin="385,27,0,0" Name="button1" VerticalAlignment="Top" Width="40" Click="button1_Click" />
+                    <TextBlock Height="23" HorizontalAlignment="Left" Margin="6,56,0,0" Name="textBlock2" Text="DHT listening port" VerticalAlignment="Top"  />
+                    <TextBox Height="23" HorizontalAlignment="Left" Margin="305,56,0,0" Name="textBox2" VerticalAlignment="Top" Width="120" Text="{Binding ListeningPort}" />
+                </Grid>
+            </TabItem>
+            <TabItem Name="ui" IsSelected="{Binding IsSelected, ElementName=ui_tree}" Height="0">
+                <Grid>
+                    <CheckBox Content="Draw graph grid" Height="16" HorizontalAlignment="Left" Margin="6,6,0,0" Name="checkBox2" VerticalAlignment="Top" IsChecked="{Binding DrawGrid}" />
+                    <TextBlock Height="23" HorizontalAlignment="Left" Margin="6,28,0,0" Name="textBlock3" Text="Download color" VerticalAlignment="Top" />
+                    <TextBlock Height="23" HorizontalAlignment="Left" Margin="6,55,0,0" Name="textBlock4" Text="Upload color" VerticalAlignment="Top" />
+                    <Button Height="23" HorizontalAlignment="Left" Margin="350,24,0,0" Name="downcolor" VerticalAlignment="Top" Width="75" Click="SelectDownloadColor" Background="{Binding DownloadBrush}" />
+                    <Button Height="23" HorizontalAlignment="Left" Margin="350,53,0,0" Name="upcolor" VerticalAlignment="Top" Width="75" Click="SelectUploadColor" Background="{Binding UploadBrush}" />
+                </Grid>
+            </TabItem>
+        </TabControl>
+    </Grid>
+</Window>
