+++ ByteFlood/UI/TorrentProperties.xaml
@@ -1,0 +1,40 @@
+﻿<!-- 
+    ByteFlood - A BitTorrent client.
+    Copyright (C) 2014 Burak Öztunç
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+    -->
+<Window x:Class="ByteFlood.TorrentProperties"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        Title="TorrentProperties" Height="276" Width="332" Loaded="Window_Loaded">
+    <Grid>
+        <GroupBox Header="General" Height="201" HorizontalAlignment="Left" Name="groupBox1" VerticalAlignment="Top" Width="310">
+            <Grid>
+                <Label Content="Maximum download speed(Kb/s): " Height="28" HorizontalAlignment="Left" Margin="10,8,0,0" Name="label1" VerticalAlignment="Top" />
+                <TextBox Height="23" HorizontalAlignment="Left" Margin="197,12,0,0" Name="maxdown" VerticalAlignment="Top" Width="88" PreviewTextInput="PreviewTextInput" />
+                <Label Content="Maximum upload speed(Kb/s): " Height="28" HorizontalAlignment="Left" Margin="26,37,0,0" Name="label2" VerticalAlignment="Top" />
+                <TextBox Height="23" HorizontalAlignment="Left" Margin="197,41,0,0" Name="maxup" VerticalAlignment="Top" Width="88" PreviewTextInput="PreviewTextInput" />
+                <Label Content="Upload slots: " Height="28" HorizontalAlignment="Left" Margin="119,67,0,0" Name="label3" VerticalAlignment="Top" />
+                <TextBox Height="23" HorizontalAlignment="Right" Margin="0,71,13,0" Name="uploadslots" VerticalAlignment="Top" Width="88" PreviewTextInput="PreviewTextInput" />
+                <Label Content="Maximum connections: " Height="28" HorizontalAlignment="Left" Margin="65,96,0,0" Name="label4" VerticalAlignment="Top" />
+                <TextBox Height="23" HorizontalAlignment="Left" Margin="197,100,0,0" Name="maxcons" VerticalAlignment="Top" Width="88" PreviewTextInput="PreviewTextInput" />
+                <CheckBox Content="Enable peer exchange" Height="16" HorizontalAlignment="Left" Margin="10,132,0,0" Name="peerex" VerticalAlignment="Top" />
+                <CheckBox Content="Enable DHT" Height="16" HorizontalAlignment="Left" Margin="10,154,0,0" Name="dht" VerticalAlignment="Top" />
+            </Grid>
+        </GroupBox>
+        <Button Content="OK" Height="23" HorizontalAlignment="Left" Margin="223,207,0,0" Name="button1" VerticalAlignment="Top" Width="75" Click="button1_Click" />
+        <Button Content="Cancel" Height="23" HorizontalAlignment="Left" Margin="142,207,0,0" Name="button2" VerticalAlignment="Top" Width="75" Click="button2_Click" />
+    </Grid>
+</Window>
